code clean up and fix codacy reported issus.
